+++ src/index_tmp.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+  <link rel="stylesheet" href="main_old.css">
+</head>
+<body>
+  <p class="account-name">-</p>
+  <button class="login">Sign In</button>
+  <script src="./main.js"></script>
+</body>
+</html>
