--- src/main_old.css
@@ -1,296 +1,0 @@
-@font-face {
-    font-family: aharoni;
-    src: url(assets/fonts/ahronbd.ttf);
-}
-
-.site-title {
-    font-size: 10vh;
-    font-family: 'aharoni', serif;
-    text-align: center;
-    color: white;
-}
-
-.login-container {
-    text-align: right;
-    font-family: 'aharoni', serif;
-    font-size: 2vh;
-    color: white;
-    position: absolute;
-    top: 20px;
-    right: 40px;
-}
-
-.my-karma {
-    font-size: 1vh;
-}
-
-.navigator-container{
-    border: 1px dashed white;
-    margin-top: -100px;
-    margin-left: 10%;
-    margin-right: 10%;
-    display: flex;
-    justify-content: center;
-    font-family: 'aharoni', serif;
-    font-size: 2vh;
-    color: white;
-    padding:1%;
-}
-
-.navigator {
-    padding:1%;
-}
-
-.bg {
-    position: fixed;
-    z-index: -10;
-    top: -50%;
-    left: -50%;
-    width: 200%;
-    height: 200%;
-}
-.bg__image {
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    margin: auto;
-    min-width: 50%;
-    min-height: 50%;
-}
-
-
-.gallery {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, 400px);
-    grid-gap: 50px;
-    justify-content: space-around;
-
-    background-color: black;
-    height: 100%;
-    margin-left: 10%;
-    margin-right: 10%;
-    margin-top: 50px;
-
-    padding: 30px;
-}
-
-
-.gallery-alert {
-    margin-left: 10%;
-    margin-right: 10%;
-    background-color: black;
-
-    min-height: calc(70vh - 50px);
-}
-
-.mint-image {
-    margin: auto;
-    width: 400px;
-    height: 400px;
-    background-color: lightgrey;
-    border: 2px solid black;
-}
-
-.modal-btn {
-    cursor: pointer;
-    border: none;
-    font-size: 1.5vh;
-    font-family: 'aharoni', serif;
-    background-color: black;
-    color: white;
-    width: 130px;
-    height: 40px;
-    margin: 10px;
-}
-
-.modal-btn:hover {
-    background-color: white;
-    color: black;
-    border: 1px solid black;
-}
-
-.modal-btn:active {
-    width: 140px;
-    height: 30px;
-    margin-top: 15px;
-    margin-bottom: 15px;
-}
-
-.modal-input {
-    outline: none;
-    border: 1px solid black;
-    background-color: white;
-    border-radius: 4px;
-}
-
-.modal-input:focus {
-    background-color: rgba(211, 211, 211, 0.12);
-}
-
-
-.btn {
-    cursor: pointer;
-}
-
-.btn:hover {
-    color: grey;
-}
-
-
-.slider {
-    -webkit-appearance: none;
-    width: 300px;
-    height: 10px;
-    border-radius: 5px;
-    background: #d3d3d3;
-    outline: none;
-    opacity: 0.7;
-    -webkit-transition: .2s;
-    transition: opacity .2s;
-}
-
-.slider::-webkit-slider-thumb {
-    -webkit-appearance: none;
-    appearance: none;
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    background: black;
-    cursor: pointer;
-}
-
-.slider::-moz-range-thumb {
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    background: #04AA6D;
-    cursor: pointer;
-}
-
-
-
-@media screen and (max-width: 1100px)  {
-    .site-title {
-        margin-top: 150px;
-    }
-
-    .mint-image {
-        width: 300px;
-        height: 300px;
-    }
-
-}
-
-.alert {
-    text-align: center;
-    padding-left: calc(50% - 200px);
-    padding-top: 100px;
-    font-family: 'aharoni', serif;
-    width: 400px;
-    height: 100px;
-    font-size: 2vh;
-    color: darkred;
-}
-
-.nft {
-    width: 400px;
-    height: 500px;
-
-    background-color: lightgrey;
-    border: 2px solid grey;
-
-    text-align: center;
-
-}
-
-
-.nft__image {
-    cursor: pointer;
-    width: 400px;
-    height: 400px;
-    border-bottom: 2px solid grey;
-}
-
-.nft__image:hover {
-    opacity: 0.5;
-}
-
-.container_image {
-    object-fit: contain;
-    width: 100%;
-    height: 100%;
-}
-
-.locked_nft {
-    margin-top: 10px;
-    height: 100px;
-    background: red;
-}
-
-
-/* The Modal (background) */
-.modal {
-    display: none; /* Hidden by default */
-    position: fixed; /* Stay in place */
-    z-index: 1; /* Sit on top */
-    padding-top: 100px; /* Location of the box */
-    left: 0;
-    top: 0;
-    width: 100%; /* Full width */
-    height: 100%; /* Full height */
-    overflow: auto; /* Enable scroll if needed */
-    background-color: rgb(0,0,0); /* Fallback color */
-    background-color: rgba(0,0,0,0.4); /* Black w/ opacity */
-}
-
-/* Modal Content */
-.modal-content {
-    text-align: center;
-    position: relative;
-    background-color: #fefefe;
-    margin: auto;
-    padding: 0;
-    border: 1px solid #888;
-    width: 60%;
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2),0 6px 20px 0 rgba(0,0,0,0.19);
-    -webkit-animation-name: animatetop;
-    -webkit-animation-duration: 0.4s;
-    animation-name: animatetop;
-    animation-duration: 0.4s
-}
-
-/* Add Animation */
-@-webkit-keyframes animatetop {
-    from {top:-300px; opacity:0}
-    to {top:0; opacity:1}
-}
-
-@keyframes animatetop {
-    from {top:-300px; opacity:0}
-    to {top:0; opacity:1}
-}
-
-/* The Close Button */
-.close {
-    color: white;
-    float: right;
-    font-size: 28px;
-    font-weight: bold;
-}
-
-.close:hover,
-.close:focus {
-    color: grey;
-    text-decoration: none;
-    cursor: pointer;
-}
-
-.modal-header {
-    padding: 2px 16px;
-    background-color: black;
-    color: white;
-}
-
-.modal-form {padding: 2px 16px;}
