+++ src/main_tmp.css
@@ -1,0 +1,4 @@
+@font-face {
+    font-family: aharoni;
+    src: url(assets/fonts/ahronbd.ttf);
+}
